+++ parser/src/error/type_not_found.rs
@@ -1,5 +1,5 @@
 use super::Error;
-use crate::ast::Ident;
+use crate::ast::{NamedRef, NamedRefKind};
 use crate::diag::{Diagnostic, DiagnosticKind, Formatted, Formatter};
 use crate::validate::Validate;
 use crate::{util, Parsed};
@@ -7,13 +7,24 @@ use crate::{util, Parsed};
 #[derive(Debug)]
 pub struct TypeNotFound {
     schema_name: String,
-    ident: Ident,
+    named_ref: NamedRef,
     candidate: Option<String>,
 }
 
 impl TypeNotFound {
-    pub(crate) fn validate(ident: &Ident, validate: &mut Validate) {
-        let schema = validate.get_current_schema();
+    pub(crate) fn validate(named_ref: &NamedRef, validate: &mut Validate) {
+        let (schema, ident) = match named_ref.kind() {
+            NamedRefKind::Intern(ident) => (validate.get_current_schema(), ident),
+
+            NamedRefKind::Extern(schema, ident) => {
+                let Some(schema) = validate.get_schema(schema.value()) else {
+                    return;
+                };
+
+                (schema, ident)
+            }
+        };
+
         for def in schema.definitions() {
             if def.name().value() == ident.value() {
                 return;
@@ -24,13 +35,13 @@ impl TypeNotFound {
 
         validate.add_error(Self {
             schema_name: validate.schema_name().to_owned(),
-            ident: ident.clone(),
+            named_ref: named_ref.clone(),
             candidate,
         });
     }
 
-    pub fn ident(&self) -> &Ident {
-        &self.ident
+    pub fn named_ref(&self) -> &NamedRef {
+        &self.named_ref
     }
 
     pub fn candidate(&self) -> Option<&str> {
@@ -48,19 +59,40 @@ impl Diagnostic for TypeNotFound {
     }
 
     fn format<'a>(&'a self, parsed: &'a Parsed) -> Formatted<'a> {
-        let mut fmt = Formatter::new(self, format!("type `{}` not found", self.ident.value()));
+        let (mut fmt, schema) = match self.named_ref.kind() {
+            NamedRefKind::Intern(ident) => (
+                Formatter::new(self, format!("type `{}` not found", ident.value())),
+                None,
+            ),
+
+            NamedRefKind::Extern(schema, ident) => (
+                Formatter::new(
+                    self,
+                    format!("type `{}::{}` not found", schema.value(), ident.value()),
+                ),
+                Some(schema),
+            ),
+        };
 
         if let Some(schema) = parsed.get_schema(&self.schema_name) {
             fmt.main_block(
                 schema,
-                self.ident.span().from,
-                self.ident.span(),
+                self.named_ref.span().from,
+                self.named_ref.span(),
                 "type used here",
             );
         }
 
         if let Some(ref candidate) = self.candidate {
-            fmt.help(format!("did you mean `{candidate}`?"));
+            match schema {
+                Some(schema) => {
+                    fmt.help(format!("did you mean `{}::{candidate}`?", schema.value()));
+                }
+
+                None => {
+                    fmt.help(format!("did you mean `{candidate}`?"));
+                }
+            }
         }
 
         fmt.format()